Split hazardous_waste.rs into hazardous_waste/ subdirectory
Split crates/simulation/src/hazardous_waste.rs (638 lines) into focused modules
under hazardous_waste/ for better maintainability:
- constants.rs (31 lines): all numeric constants
- types.rs (80 lines): TreatmentType enum + HazardousWasteState resource
- systems.rs (166 lines): update system + helper functions
- tests.rs (339 lines): all unit tests
- mod.rs (37 lines): thin facade with pub use re-exports + plugin

All public items re-exported from mod.rs — no external import changes needed.
