+++ crates/simulation/src/hazardous_waste/tests.rs
@@ -1,297 +1,11 @@
-//! Hazardous waste facility and industrial waste management (WASTE-007).
-//!
-//! Industrial and medical buildings generate hazardous waste that requires
-//! specialized treatment at a HazardousWasteFacility. Without adequate
-//! treatment capacity, overflow waste triggers illegal dumping events that
-//! cause soil and groundwater contamination, plus federal fines.
-//!
-//! Key design points:
-//! - HazardousWasteFacility: 20 tons/day capacity, $3M build, $5K/day operating
-//! - Industrial buildings generate waste based on level (higher level = more waste)
-//! - Medical buildings (Hospital, MedicalClinic, MedicalCenter) generate medical waste
-//! - Without facility: illegal dumping causes contamination + federal fines
-//! - Treatment types: chemical, thermal, biological, stabilization
-
-use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::buildings::Building;
-use crate::grid::ZoneType;
-use crate::groundwater::WaterQualityGrid;
-use crate::services::{ServiceBuilding, ServiceType};
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-/// Capacity per hazardous waste facility in tons/day.
-pub const FACILITY_CAPACITY_TONS_PER_DAY: f32 = 20.0;
-
-/// Build cost for a single hazardous waste facility ($3M).
-pub const FACILITY_BUILD_COST: f64 = 3_000_000.0;
-
-/// Operating cost per facility per day ($5K).
-pub const FACILITY_OPERATING_COST_PER_DAY: f64 = 5_000.0;
-
-/// Federal fine per illegal dump event ($50K).
-pub const FEDERAL_FINE_PER_EVENT: f64 = 50_000.0;
-
-/// Groundwater quality reduction per unit of illegal dumping overflow.
-/// Applied to cells around industrial buildings when dumping occurs.
-pub const CONTAMINATION_PER_OVERFLOW_TON: f32 = 2.0;
-
-/// Contamination natural decay rate per slow tick (1% reduction).
-pub const CONTAMINATION_DECAY_RATE: f32 = 0.01;
-
-/// Base hazardous waste generation rate per industrial building level (tons/day).
-/// Level 1 = 0.5, Level 2 = 1.0, Level 3 = 2.0, Level 4 = 3.5, Level 5 = 5.0.
-pub const INDUSTRIAL_WASTE_PER_LEVEL: [f32; 5] = [0.5, 1.0, 2.0, 3.5, 5.0];
-
-/// Base hazardous waste generation for medical buildings (tons/day per facility).
-pub const MEDICAL_WASTE_RATE: f32 = 0.8;
-
-/// Radius (in grid cells) around industrial buildings affected by illegal dumping
-/// contamination of groundwater.
-pub const CONTAMINATION_RADIUS: i32 = 4;
-
-// =============================================================================
-// Treatment type enum
-// =============================================================================
-
-/// Treatment method used by a hazardous waste facility.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub enum TreatmentType {
-    /// Chemical treatment: neutralization, oxidation, reduction.
-    #[default]
-    Chemical,
-    /// Thermal treatment: incineration at high temperatures.
-    Thermal,
-    /// Biological treatment: bioremediation using microorganisms.
-    Biological,
-    /// Stabilization/solidification: encapsulating waste in concrete or polymers.
-    Stabilization,
-}
-
-impl TreatmentType {
-    /// Efficiency multiplier for treatment capacity.
-    /// Higher efficiency means more waste can be treated per unit capacity.
-    pub fn efficiency(&self) -> f32 {
-        match self {
-            TreatmentType::Chemical => 1.0,
-            TreatmentType::Thermal => 1.2,
-            TreatmentType::Biological => 0.8,
-            TreatmentType::Stabilization => 0.9,
-        }
-    }
-
-    /// Cost multiplier relative to base operating cost.
-    pub fn cost_multiplier(&self) -> f64 {
-        match self {
-            TreatmentType::Chemical => 1.0,
-            TreatmentType::Thermal => 1.5,
-            TreatmentType::Biological => 0.7,
-            TreatmentType::Stabilization => 0.8,
-        }
-    }
-}
-
-// =============================================================================
-// HazardousWasteState resource
-// =============================================================================
-
-/// City-wide hazardous waste management state.
-///
-/// Tracks generation, treatment capacity, overflow, contamination, and fines.
-#[derive(Resource, Default, Clone, Debug, Serialize, Deserialize)]
-pub struct HazardousWasteState {
-    /// Total hazardous waste generated per day (tons/day).
-    pub total_generation: f32,
-    /// Total treatment capacity available (tons/day).
-    pub treatment_capacity: f32,
-    /// Untreated waste overflow (tons) — resets each tick but accumulates effects.
-    pub overflow: f32,
-    /// Cumulative count of illegal dump events.
-    pub illegal_dump_events: u32,
-    /// Accumulated ground contamination level (0.0 = clean).
-    pub contamination_level: f32,
-    /// Accumulated federal fines in dollars.
-    pub federal_fines: f64,
-    /// Number of active hazardous waste treatment facilities.
-    pub facility_count: u32,
-    /// Daily operating cost for all facilities.
-    pub daily_operating_cost: f64,
-    /// Breakdown of waste by treatment type (tons processed per type).
-    pub chemical_treated: f32,
-    /// Tons treated via thermal methods.
-    pub thermal_treated: f32,
-    /// Tons treated via biological methods.
-    pub biological_treated: f32,
-    /// Tons treated via stabilization methods.
-    pub stabilization_treated: f32,
-}
-
-// =============================================================================
-// Helper functions
-// =============================================================================
-
-/// Calculate hazardous waste generation for an industrial building based on level.
-pub fn industrial_waste_generation(level: u8) -> f32 {
-    let idx = (level as usize)
-        .saturating_sub(1)
-        .min(INDUSTRIAL_WASTE_PER_LEVEL.len() - 1);
-    INDUSTRIAL_WASTE_PER_LEVEL[idx]
-}
-
-/// Calculate the number of hazardous waste treatment facilities from service buildings.
-/// Uses Incinerator as the closest existing ServiceType for hazardous waste treatment,
-/// since a dedicated HazardousWasteFacility ServiceType doesn't exist yet.
-pub fn count_hazardous_facilities(services: &[&ServiceBuilding]) -> u32 {
-    services
-        .iter()
-        .filter(|s| s.service_type == ServiceType::Incinerator)
-        .count() as u32
-}
-
-/// Determine if a service building qualifies as a medical facility that generates
-/// hazardous waste.
-pub fn is_medical_waste_generator(service_type: ServiceType) -> bool {
-    matches!(
-        service_type,
-        ServiceType::Hospital | ServiceType::MedicalClinic | ServiceType::MedicalCenter
-    )
-}
-
-// =============================================================================
-// System
-// =============================================================================
-
-/// Updates hazardous waste state each slow tick.
-///
-/// 1. Tallies hazardous waste generation from industrial + medical buildings
-/// 2. Counts treatment facilities (Incinerators serving as hazardous waste facilities)
-/// 3. Computes overflow (generation - capacity)
-/// 4. If overflow > 0: triggers illegal dumping, contaminates groundwater, adds fines
-/// 5. Applies natural contamination decay
-#[allow(clippy::too_many_arguments)]
-pub fn update_hazardous_waste(
-    slow_timer: Res<crate::SlowTickTimer>,
-    mut state: ResMut<HazardousWasteState>,
-    buildings: Query<&Building>,
-    services: Query<&ServiceBuilding>,
-    mut water_quality: Option<ResMut<WaterQualityGrid>>,
-) {
-    if !slow_timer.should_run() {
-        return;
-    }
-
-    // --- Phase 1: Calculate total hazardous waste generation ---
-    let mut total_generation: f32 = 0.0;
-
-    // Industrial buildings generate hazardous waste based on level
-    for building in &buildings {
-        if building.zone_type == ZoneType::Industrial {
-            total_generation += industrial_waste_generation(building.level);
-        }
-    }
-
-    // Medical facilities generate medical hazardous waste
-    for service in &services {
-        if is_medical_waste_generator(service.service_type) {
-            total_generation += MEDICAL_WASTE_RATE;
-        }
-    }
-
-    state.total_generation = total_generation;
-
-    // --- Phase 2: Count treatment facilities and compute capacity ---
-    // Use Incinerators as proxy for hazardous waste treatment facilities
-    let mut facility_count: u32 = 0;
-    for service in &services {
-        if service.service_type == ServiceType::Incinerator {
-            facility_count += 1;
-        }
-    }
-
-    state.facility_count = facility_count;
-    state.treatment_capacity = facility_count as f32 * FACILITY_CAPACITY_TONS_PER_DAY;
-    state.daily_operating_cost = facility_count as f64 * FACILITY_OPERATING_COST_PER_DAY;
-
-    // --- Phase 3: Compute overflow ---
-    let overflow = (total_generation - state.treatment_capacity).max(0.0);
-    state.overflow = overflow;
-
-    // --- Phase 4: Distribute treated waste across treatment types ---
-    // For now, all facilities use chemical treatment by default.
-    // In a future PR, individual facility treatment types can be tracked.
-    let treated = total_generation.min(state.treatment_capacity);
-    state.chemical_treated = treated;
-    state.thermal_treated = 0.0;
-    state.biological_treated = 0.0;
-    state.stabilization_treated = 0.0;
-
-    // --- Phase 5: Handle overflow — illegal dumping ---
-    if overflow > 0.0 {
-        state.illegal_dump_events += 1;
-        state.federal_fines += FEDERAL_FINE_PER_EVENT;
-
-        // Contamination accumulates based on overflow amount
-        let contamination_increase = overflow * CONTAMINATION_PER_OVERFLOW_TON;
-        state.contamination_level += contamination_increase;
-
-        // Apply groundwater contamination around industrial buildings
-        if let Some(ref mut wq) = water_quality {
-            for building in &buildings {
-                if building.zone_type != ZoneType::Industrial {
-                    continue;
-                }
-                let cx = building.grid_x as i32;
-                let cy = building.grid_y as i32;
-                // Scale contamination by overflow magnitude (capped)
-                let quality_reduction = (overflow * 0.5).min(10.0) as u8;
-                if quality_reduction == 0 {
-                    continue;
-                }
-                for dy in -CONTAMINATION_RADIUS..=CONTAMINATION_RADIUS {
-                    for dx in -CONTAMINATION_RADIUS..=CONTAMINATION_RADIUS {
-                        let nx = cx + dx;
-                        let ny = cy + dy;
-                        if nx < 0
-                            || ny < 0
-                            || (nx as usize) >= wq.width
-                            || (ny as usize) >= wq.height
-                        {
-                            continue;
-                        }
-                        let dist = dx.abs() + dy.abs();
-                        let falloff = quality_reduction.saturating_sub(dist as u8);
-                        if falloff > 0 {
-                            let ux = nx as usize;
-                            let uy = ny as usize;
-                            let idx = uy * wq.width + ux;
-                            wq.levels[idx] = wq.levels[idx].saturating_sub(falloff);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // --- Phase 6: Natural contamination decay ---
-    if state.contamination_level > 0.0 {
-        state.contamination_level *= 1.0 - CONTAMINATION_DECAY_RATE;
-        if state.contamination_level < 0.01 {
-            state.contamination_level = 0.0;
-        }
-    }
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
+//! Tests for hazardous waste management.
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use crate::hazardous_waste::constants::*;
+    use crate::hazardous_waste::systems::*;
+    use crate::hazardous_waste::types::*;
+    use crate::services::{ServiceBuilding, ServiceType};
 
     // -------------------------------------------------------------------------
     // TreatmentType tests
@@ -621,18 +335,5 @@ mod tests {
         assert!(TreatmentType::Thermal.efficiency() > TreatmentType::Chemical.efficiency());
         // Biological should be least efficient
         assert!(TreatmentType::Biological.efficiency() < TreatmentType::Chemical.efficiency());
-    }
-}
-
-pub struct HazardousWastePlugin;
-
-impl Plugin for HazardousWastePlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<HazardousWasteState>().add_systems(
-            FixedUpdate,
-            update_hazardous_waste
-                .after(crate::imports_exports::process_trade)
-                .in_set(crate::SimulationSet::Simulation),
-        );
     }
 }
